requirements.txt
fastapi
uvicorn[standard]
pyyaml
requests
pydantic
flask
